Migrate to .NET 10, delete NList<T> and NListHashSet<T>, add BaseMutableIndexable<T>, fix bugs in slices

+++ Tests/NStar.ExtraLibs.Tests/ListsExtraTests.cs
@@ -658,10 +658,10 @@ public class BaseSumListTests<T, TCertain> where T : INumber<T> where TCertain :
 			Assert.AreEqual(sl2, sl);
 		}, () =>
 		{
-			var length = Min(random.Next(17), (int)sl.Length);
+			var length = Min(random.Next(17), sl.Length);
 			if (sl.Length < length)
 				return;
-			var start = random.Next((int)sl.Length - length + 1);
+			var start = random.Next(sl.Length - length + 1);
 			var sl2 = sl.Reverse(start, length);
 			gl.Reverse(start, length);
 			Assert.IsTrue(sl.Equals(gl));
